Creating and viewing a BTS + GUI
An application that reads integers and places them in a BST.
The end of inputting the numbers constructing the tree is done by entering the value -999.
The user then enters the integer values ​​that he wants to be deleted from the tree. End of inputting numbers to be deleted from BST
is done by entering the value -999. When there is a number to be deleted that is not in the
tree, the program ignores it.